Using BigDecimal for money amounts Allowing FREETEXT comm between pythonanywhere server and app

+++ app/src/main/java/com/gb/canibuyit/db/model/ApiSaving.kt
@@ -17,5 +17,5 @@ class ApiSaving(
     @DatabaseField(columnName = Contract.Savings.CREATED, canBeNull = false, dataType = DataType.SERIALIZABLE)
     var created: LocalDate? = null,
     @DatabaseField(columnName = Contract.Savings.TARGET, canBeNull = false)
-    var target: Double? = null
+    var target: Int? = null
 )
